+++ packages/generator-typescript/templates/gjs/module-import.js
@@ -1,0 +1,2 @@
+const gi = globalThis.imports?.gi || {};
+export default gi;